Remove risk of getting "coroutine 'Queue.get' was never awaited" warning.
This could happen when the coroutine object returned from
asyncio.Queue.get() was handed over to a newly created task which was then
immediately cancelled so that the coroutine object was never awaited.

Instead, let the newly created task itself make the call to
asyncio.Queue.get() before immediately awaiting it.

Resolves #646.

Signed-off-by: Knut Aksel Røysland <knutroy@ifi.uio.no>